Update questionnaires.js
attempted to add in the Very Short Authoritarianism scale, which is a likert-scale 6-question survey. This is clearly not finished but i'm hoping it's on the right track. Need to figure out how to work with reverse scoring on certain questions, and how I can make a temporary timeline that I can then quickly test to see if it's doing what I intend